refactor: extract QR sign-in state views into feature widgets
QrScannerView, QrSigninVerifyingView and QrSigninPermissionView move
from private classes inside the screen to the feature's widgets/
directory (mirroring the other features); the screen keeps only the
scanner lifecycle and detection handling.
