+++ lib/features/login/features/login_qr_signin/login_qr_signin.dart
@@ -1,3 +1,4 @@
 export 'cubit/cubit.dart';
 export 'models/models.dart';
 export 'view/view.dart';
+export 'widgets/widgets.dart';
