+++ lib/features/login/features/login_qr_signin/widgets/widgets.dart
@@ -1,0 +1,3 @@
+export 'qr_scanner_view.dart';
+export 'qr_signin_permission_view.dart';
+export 'qr_signin_verifying_view.dart';
